web-backend/requirements.txt
# Requirements for ResDB
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
anyio==3.6.2
ariadne==0.17.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptoconditions==0.8.1
cryptography==3.4.7
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
graphql-core==3.2.3
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==0.4.9
MarkupSafe==2.1.1
multidict==6.0.4
mypy-extensions==0.4.3
pyasn1==0.4.8
pycparser==2.21
Pygments==2.14.0
PyNaCl==1.4.0
pysha3==1.0.2
python-dateutil==2.8.2
python-multipart==0.0.5
python-rapidjson==1.8
PyYAML==6.0
requests==2.32.3
rich==13.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.41.3
strawberry-graphql==0.152.0
typing-inspect==0.8.0
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0

# Requirements for application
PyJWT==2.10.1
pymongo==4.10.1
fastapi==0.115.6
python-dotenv==1.0.1
boto3==1.35.76
bcrypt==4.2.1
pillow==11.0.0
PyPDF2==3.0.1
reportlab==4.2.5
pyHanko[pkcs11,image-support,opentype,xmp]==0.20.1

